requirements.txt
pandas
numpy
scipy
scikit-learn
lightgbm
xgboost
sqlalchemy
psycopg2-binary
requests
python-dotenv
prefect
matplotlib
seaborn
telegram-send
mplfinance
ta
python-telegram-bot
asyncio
aiohttp
websockets
pydantic-settings
tenacity
streamlit
plotly
